backend/requirements.txt
flask==3.1.1
flask-cors==5.0.1
python-dotenv==1.1.0
temporalio==1.11.0
